# frontend/public/data/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Anuvaad API",
    "description": "**Anuvaad** \u2014 AI-powered code translation platform.\n\nTranslate code to plain English, English to code, and between 35+ languages.\n\nAll new clients should use the `/api/v1/` prefix. Legacy `/api/` routes are deprecated and will be removed on **2027-01-01**.",
    "contact": {
      "name": "Anuvaad Support",
      "url": "http://localhost:3000/"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "3.0.0"
  },
  "paths": {
...
    "/api/v1/create-checkout-session": {
      "post": {
        "tags": [
          "billing"
        ],
        "summary": "Create Checkout Session",
        "description": "Create a subscription checkout session (Stripe primary, Razorpay fallback).",
        "operationId": "create_checkout_session_api_v1_create_checkout_session_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/billing/create-checkout-session": {
      "post": {
        "tags": [
          "billing"
        ],
        "summary": "Create Checkout Session",
        "description": "Create a subscription checkout session (Stripe primary, Razorpay fallback).",
        "operationId": "create_checkout_session_api_v1_billing_create_checkout_session_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/create-portal-session": {
      "post": {
        "tags": [
          "billing"
        ],
        "summary": "Create Portal Session",
        "description": "Generate a self-service customer portal session for subscription management.",
        "operationId": "create_portal_session_api_v1_create_portal_session_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/billing/create-portal-session": {
      "post": {
        "tags": [
          "billing"
        ],
        "summary": "Create Portal Session",
        "description": "Generate a self-service customer portal session for subscription management.",
        "operationId": "create_portal_session_api_v1_billing_create_portal_session_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/create-credit-checkout": {
      "post": {
        "tags": [
          "billing"
        ],
        "summary": "Create Credit Checkout",
        "description": "Create a Razorpay one-time order for buying 100 translation credits (\u20b9100).",
        "operationId": "create_credit_checkout_api_v1_create_credit_checkout_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/verify-payment": {
      "post": {
        "tags": [
          "billing"
        ],
        "summary": "Verify Payment",
        "description": "Verify Razorpay HMAC signature then activate Pro or top up credits.\nDelegates all business logic to BillingService.",
        "operationId": "verify_payment_api_v1_verify_payment_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyPaymentPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/subscription-status": {
      "get": {
        "tags": [
          "billing"
        ],
        "summary": "Get Subscription Status",
        "description": "Return the user's current subscription plan.\nGET endpoint for SWR caching (P4: replaces the old POST version).",
        "operationId": "get_subscription_status_api_v1_subscription_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/check-credits": {
      "get": {
        "tags": [
          "billing"
        ],
        "summary": "Get Check Credits",
        "description": "Return the user's current translation credit balance and remaining daily quota.",
        "operationId": "get_check_credits_api_v1_check_credits_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/create-checkout-session": {
      "post": {
        "tags": [
          "billing"
        ],
        "summary": "Create Checkout Session",
        "description": "Create a subscription checkout session (Stripe primary, Razorpay fallback).",
        "operationId": "create_checkout_session_api_create_checkout_session_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/billing/create-checkout-session": {
      "post": {
        "tags": [
          "billing"
        ],
        "summary": "Create Checkout Session",
        "description": "Create a subscription checkout session (Stripe primary, Razorpay fallback).",
        "operationId": "create_checkout_session_api_billing_create_checkout_session_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/create-portal-session": {
      "post": {
        "tags": [
          "billing"
        ],
        "summary": "Create Portal Session",
        "description": "Generate a self-service customer portal session for subscription management.",
        "operationId": "create_portal_session_api_create_portal_session_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/billing/create-portal-session": {
      "post": {
        "tags": [
          "billing"
        ],
        "summary": "Create Portal Session",
        "description": "Generate a self-service customer portal session for subscription management.",
        "operationId": "create_portal_session_api_billing_create_portal_session_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/create-credit-checkout": {
      "post": {
        "tags": [
          "billing"
        ],
        "summary": "Create Credit Checkout",
        "description": "Create a Razorpay one-time order for buying 100 translation credits (\u20b9100).",
        "operationId": "create_credit_checkout_api_create_credit_checkout_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/verify-payment": {
      "post": {
        "tags": [
          "billing"
        ],
        "summary": "Verify Payment",
        "description": "Verify Razorpay HMAC signature then activate Pro or top up credits.\nDelegates all business logic to BillingService.",
        "operationId": "verify_payment_api_verify_payment_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyPaymentPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/subscription-status": {
      "get": {
        "tags": [
          "billing"
        ],
        "summary": "Get Subscription Status",
        "description": "Return the user's current subscription plan.\nGET endpoint for SWR caching (P4: replaces the old POST version).",
        "operationId": "get_subscription_status_api_subscription_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/check-credits": {
      "get": {
        "tags": [
          "billing"
        ],
        "summary": "Get Check Credits",
        "description": "Return the user's current translation credit balance and remaining daily quota.",
        "operationId": "get_check_credits_api_check_credits_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CheckoutPayload": {
        "properties": {
          "user_email": {
            "type": "string",
            "maxLength": 254,
            "minLength": 5,
            "title": "User Email"
          }
        },
        "type": "object",
        "required": [
          "user_email"
        ],
        "title": "CheckoutPayload",
        "description": "BACK-06: access_token removed \u2014 auth via Authorization header (Depends)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}